refactor: 코스 presentation DTO 패키지 분리
